backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
email-validator>=2.1.0
aiofiles==23.2.1
httpx==0.25.2
openai
groq
duckduckgo_search
